requirements.txt
argcomplete==3.6.2
astroid==4.0.2
asttokens==3.0.0
attrs==25.4.0
autocommand==2.2.2
bitarray==3.8.0
bitstring==4.3.1
btrfsutil==6.17.1
cachetools==6.2.2
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
cupshelpers==1.0
dbus-python==1.4.0
dill==0.4.0
distro==1.9.0
docutils==0.22.3
ecdsa==0.19.1
esptool==5.1.0
fastjsonschema==2.21.2
filelock==3.20.0
gps==3.27
hyprshade==4.0.0
idna==3.11
importlib_metadata==8.7.0
inputs==0.5
intelhex==2.3.0
isort==7.0.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.2
lark==1.3.1
lensfun==0.3.4
lit==21.1.6.dev0
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.8.0
mpremote==1.26.1
mypy==1.18.2
mypy_extensions==1.1.0
OpenTimelineIO==0.18.1
orjson==3.11.4
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pillow==12.0.0
pipx==1.8.0
platformdirs==4.5.0
poetry-core==2.2.1
ProtonUp-Qt==2.14.0
psutil==7.1.3
ptyprocess==0.7.0
pwquality==1.4.5
pyaaf2==1.7.1
pyaml==24.12.0
pycairo==1.29.0
pycparser==2.23
pycryptodomex==3.23.0
pycups==2.0.4
Pygments==2.19.2
PyGObject==3.54.5
pylint==4.0.3
pyserial==3.5
PySide6==6.10.0
python-dateutil==2.9.0
pyxdg==0.28
PyYAML==6.0.3
reedsolo==1.7.0
requests==2.32.5
rich==14.2.0
rich-click==1.9.4
Send2Trash==1.8.3
setuptools==80.9.0
shiboken6==6.10.0
shiboken6-generator==6.10.0
shtab==1.8.0
six==1.17.0
steam==1.6.1
termcolor==3.2.0
terminaltexteffects==0.13.0
thonny==4.1.7
tldr==3.4.3
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.15.0
ufw==0.36.2
urllib3==2.5.0
userpath==1.9.2
vdf==4.0
visidata==3.3
wheel==0.45.1
zipp==3.21.0
zstandard==0.25.0
